+++ ntex-http/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [0.1.0] - 2022-xx-xx
 
+* impl TryFrom<&str> for Value
+
 * impl FromIterator for HeaderMap
 
 * Move http types to separate crate